requirements.txt
gradio
torch
torchvision
torchaudio
pillow